Reorganize modules & imports
