+++ fictionsuit/api_wrap/user_message.py
@@ -1,5 +1,6 @@
 from abc import ABC, abstractmethod
-from ..core.core import OpenAIChat
+
+from ..core import OpenAIChat
 
 
 class UserMessage(ABC):
