+++ fictionsuit/commands/__init__.py
@@ -1,0 +1,17 @@
+from .chat import Chat
+from .command_group import (
+    CommandFailure,
+    CommandGroup,
+    CommandHandled,
+    CommandNotFound,
+    CommandNotHandled,
+    CommandReply,
+    CommandResult,
+    auto_reply,
+    command_split,
+    multi_command,
+    slow_command,
+)
+from .debug import Debug
+from .research import Research
+from .scripting import Scripting
