+++ fictionsuit/commands/command_group.py
@@ -1,9 +1,10 @@
-import traceback
-import inspect
 import string
-from typing import Any, Callable, NewType, TypeVar
+import traceback
+from typing import Any, Callable
 
-from ..api_wrap.user_message import UserMessage
+from ..api_wrap import UserMessage
+
+# TODO: this file is doing way too many things. probably best to extract to a module
 
 
 class CommandNotFound:
