+++ fictionsuit/commands/research.py
@@ -1,7 +1,8 @@
+from ..api_wrap import UserMessage
+from ..core import summarize
+from ..db.supa import list_first_article, supabase
+from ..utils import scrape_link
 from .command_group import CommandGroup
-from ..core.core import summarize, scrape_link
-from ..api_wrap.user_message import UserMessage
-from ..db.supa import supabase, list_first_article
 
 
 class Research(CommandGroup):
