+++ fictionsuit/core/fictionscript/__init__.py
@@ -1,0 +1,3 @@
+from .fictionscript import FictionScript
+from .script_message import ScriptMessage
+from .varscope import VarScope
